# Conflicts:
#	src/main/resources/db/data.sql

+++ src/main/java/org/example/metabox/review/Review.java
@@ -27,13 +27,13 @@ public class Review {
 
     private String comment;
     // TODO : Enum 처리?
-    private double rating;
+    private int rating;
 
     @CreationTimestamp
     private LocalDateTime createdAt;    // 댓글 작성 시간
 
     @Builder
-    public Review(int id, User user, Movie movie, String comment, double rating) {
+    public Review(int id, User user, Movie movie, String comment, int rating) {
         this.id = id;
         this.user = user;
         this.movie = movie;
